Added the "sound" interface to player. It takes a single command with an index argument (short) that specifies the sound file to play. I added a driver, p2os_sound, that implements the sound interface. This lets you access the amigobot sounds. It is now device 27 in player.h.
This driver is a subdriver of p2os, although currently only the amigo
bot supports it. It will be linked in by default with the p2os drivers

I have updated the Javaclient client to support this interface with
a class called SoundPlayerDevice.

I fixed a marshalling bug in p2os Main that casted a char pointer
into a short pointer, something that isn't kosher on the ARM because
of even byte alignment. There were expanded into single byte
assignments using | and bit shifting.

Fixed a bug in the p2os set odometry configuration command. Was simply
adding the offset to the current x, y and heading from the robot. Changed
this to do true coordinate frame rotation when heading offset is not zero.

Tested resulting binary on both i686 and arm.